+++ __tests__/hostApi/statementStore.spec.ts
@@ -108,7 +108,7 @@ describe('Host API: StatementStore', () => {
 
     const result = await statementStore.createProof(accountId, statement);
 
-    expect(handler).toBeCalledWith([mockWireAccountId, statement], {
+    expect(handler).toHaveBeenCalledWith([mockWireAccountId, statement], {
       ok: expect.any(Function),
       err: expect.any(Function),
     });
@@ -130,7 +130,7 @@ describe('Host API: StatementStore', () => {
 
     await statementStore.submit(signedStatement);
 
-    expect(handler).toBeCalledWith(signedStatement, { ok: expect.any(Function), err: expect.any(Function) });
+    expect(handler).toHaveBeenCalledWith(signedStatement, { ok: expect.any(Function), err: expect.any(Function) });
     expect(permissionHandler).toHaveBeenCalledOnce();
     const [receivedParams] = permissionHandler.mock.calls[0]!;
     expect(receivedParams).toEqual({ tag: 'StatementSubmit', value: undefined });